Rename theme context
