+++ TabbedTSApp/src/screens/MasterDetail/ItemDetail/ItemDetailScreen.tsx
@@ -3,7 +3,7 @@ import {View, Text, ScrollView, SafeAreaView} from 'react-native';
 import Icon from 'react-native-vector-icons/Ionicons';
 
 import {getStyles} from './ItemDetailScreen.style';
-import {useTheme} from '../../../context/Theme.context';
+import {useTheme} from '../../../context/ThemeContext';
 
 export interface IProps {
   item: any;
