+++ TabbedTSApp/windows/TabbedTSApp/AutolinkedNativeModules.g.targets
@@ -2,5 +2,9 @@
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <!-- AutolinkedNativeModules.g.targets contents generated by "react-native autolink-windows" -->
   <ItemGroup>
+    <!-- Projects from @react-native-async-storage/async-storage -->
+    <ProjectReference Include="$(ProjectDir)..\..\node_modules\@react-native-async-storage\async-storage\windows\ReactNativeAsyncStorage\ReactNativeAsyncStorage.vcxproj">
+      <Project>{4855D892-E16C-404D-8286-0089E0F7F9C4}</Project>
+    </ProjectReference>
   </ItemGroup>
 </Project>
